backend/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiomysql==0.2.0
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# Excel/CSV 处理
pandas==2.2.0
openpyxl==3.1.2

# Word 文档生成
python-docx==1.1.0

# 工具
python-dotenv==1.0.0
aiofiles==23.2.1

# 日期时间处理
python-dateutil==2.8.2

# CORS
starlette==0.35.1

# JWT 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
